+++ resources/views/people/show.blade.php
@@ -173,14 +173,10 @@
             <div class="card mb-4">
                 <div class="card-header">
                     Transactions
-                    @if ( $transactions->perPage() < $transactions->total() )
-                        <small class="pull-right text-muted">showing latest {{ $transactions->perPage() }} transactions</small>
-                    @endif
                 </div>
-                <div class="card-body @if( ! $transactions->isEmpty() )p-0 @endif">
-
+                <div class="card-body">
                     @if( ! $transactions->isEmpty() )
-                        <table class="table table-sm table-hover m-0">
+                        <table class="table table-sm table-hover">
                             <thead>
                                 <tr>
                                     <th style="width: 200px">Date</th>
@@ -196,6 +192,7 @@
                                 @endforeach
                             </tbody>
                         </table>
+                        {{ $transactions->links('vendor.pagination.bootstrap-4') }}
                     @else
                         <div class="alert alert-info m-0">
                             No transactions found.